# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mandaean Flutter App API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friends": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Get all friend suggestions for the authenticated user",
                "operationId": "e973365285d0872598256a7d417c5047",
                "responses": {
                    "200": {
                        "description": "List of sent and received friend suggestions."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "Send a friend suggestion",
                "operationId": "baaf7ddb1837a40545487039703416bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "suggested_friend_id"
                                ],
                                "properties": {
                                    "suggested_friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Friend suggestion sent."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/{friendSuggestion}": {
            "put": {
                "tags": [
                    "Community"
                ],
                "summary": "Update the status of a friend suggestion",
                "operationId": "d7ff791aea669b977e5154a60ef71596",
                "parameters": [
                    {
                        "name": "friendSuggestion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "accepted",
                                            "rejected",
                                            "suggested"
                                        ],
                                        "example": "accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Friend suggestion updated."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Community"
                ],
                "summary": "Delete a friend suggestion or relationship",
                "operationId": "709bb1ee891d193274275153c4c92499",
                "parameters": [
                    {
                        "name": "friendSuggestion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend suggestion deleted."
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}